refactor: :recycle: Reorganize configurations
