+++ src/pages/Processes/hooks/useProcessesData.ts
@@ -4,7 +4,8 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { Role } from '../../../API/REST.enum';
 import { RESTApi } from '../../../API/REST.resources';
-import { BIG_PAGINATION_SIZE, UPDATE_INTERVAL } from '../../../config/app';
+import { BIG_PAGINATION_SIZE } from '../../../config/app';
+import { UPDATE_INTERVAL } from '../../../config/reactQuery';
 import { QueryFilters } from '../../../types/REST.interfaces';
 import { QueriesProcesses } from '../Processes.enum';
