+++ src/pages/Services/hooks/useServicesData.ts
@@ -3,7 +3,7 @@ import { startTransition, useCallback, useState } from 'react';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { RESTApi } from '../../../API/REST.resources';
-import { UPDATE_INTERVAL } from '../../../config/app';
+import { UPDATE_INTERVAL } from '../../../config/reactQuery';
 import { QueryFilters } from '../../../types/REST.interfaces';
 import { QueriesServices } from '../Services.enum';
