+++ src/pages/Topology/components/TopologyProcesses.tsx
@@ -19,7 +19,6 @@ import { GraphEdge, GraphNode, SkGraphProps } from 'types/Graph.interfaces';
 
 import TopologyDetails, { TopoloyDetailsProps } from './TopologyDetails';
 import TopologyToolbar from './TopologyToolbar';
-import { MAX_DRAWER_WIDTH, MIN_DRAWER_WIDTH } from '../../../config/app';
 import SkGraph from '../../../core/components/SkGraph';
 import { ProcessesLabels, ProcessesRoutesPaths } from '../../Processes/Processes.enum';
 import useTopologyProcessData from '../hooks/useTopologyProcessData';
@@ -37,6 +36,9 @@ import {
   displayOptionsForProcesses
 } from '../Topology.constants';
 import { TopologyLabels } from '../Topology.enum';
+
+export const MIN_DRAWER_WIDTH = 500;
+export const MAX_DRAWER_WIDTH = 800;
 
 const TopologyProcesses: FC<{
   serviceIds?: string[];
