+++ src/pages/Topology/hooks/useTopologySiteData.ts
@@ -1,8 +1,8 @@
 import { useSuspenseQueries } from '@tanstack/react-query';
 
 import { RESTApi } from '../../../API/REST.resources';
-import { UPDATE_INTERVAL } from '../../../config/app';
 import { PrometheusLabelsV2 } from '../../../config/prometheus';
+import { UPDATE_INTERVAL } from '../../../config/reactQuery';
 import { QueriesSites } from '../../Sites/Sites.enum';
 import { TopologyController } from '../services';
 import { QueriesPairs } from '../Topology.enum';
