+++ src/pages/Topology/services/index.ts
@@ -2,7 +2,6 @@ import { GraphEdge, GraphCombo, GraphNode, GraphElementNames } from 'types/Graph
 
 import { PrometheusApi } from '../../../API/Prometheus.api';
 import { Direction } from '../../../API/REST.enum';
-import { IDS_GROUP_SEPARATOR, IDS_MULTIPLE_SELECTION_SEPARATOR, PAIR_SEPARATOR } from '../../../config/app';
 import { PrometheusLabelsV2 } from '../../../config/prometheus';
 import { formatByteRate, formatBytes } from '../../../core/utils/formatBytes';
 import { removeDuplicatesFromArrayOfObjects } from '../../../core/utils/removeDuplicatesFromArrayOfObjects';
@@ -15,6 +14,10 @@ import {
   TopologyConfigMetricsParams
 } from '../../../types/Topology.interfaces';
 import { shape } from '../Topology.constants';
+
+export const IDS_GROUP_SEPARATOR = '~';
+export const IDS_MULTIPLE_SELECTION_SEPARATOR = ',';
+export const PAIR_SEPARATOR = 'processpair-';
 
 export const TopologyController = {
   getCombosFromNodes: (nodes: GraphNode[]): GraphCombo[] => {
